Moving stringify into a Utils module for the MapGenerators
