+++ lib/dungeon_crawl/map_generators/test_rooms.ex
@@ -7,7 +7,7 @@ defmodule DungeonCrawl.MapGenerators.TestRooms do
 #################   
 #...............#   
 #...............#   
-#@..............#   
+#...............#   
 ######'##########   
     #..........#    
     #..........#
