+++ test/dungeon_crawl/map_generators/empty_test.exs
@@ -1,10 +1,10 @@
-defmodule DungeonCrawl.MapGenerators.ConnectedRoomsTest do
+defmodule DungeonCrawl.MapGenerators.EmptyTest do
   use DungeonCrawl.DataCase
 
-  alias DungeonCrawl.MapGenerators.ConnectedRooms
+  alias DungeonCrawl.MapGenerators.Empty
 
   test "generate returns a map with a rol, col tuple as key and tile as value" do
-    dungeon_map = ConnectedRooms.generate(20,20)
+    dungeon_map = Empty.generate(20,20)
     [first_key | _other_keys] = Map.keys dungeon_map
 
     assert is_map(dungeon_map)
